Se agregaron comentarios y algunas optimizaciones de código

+++ include/HC-SR04.h
@@ -1,15 +1,16 @@
 #ifndef HCSR04
 #define HCSR04
 
+#include <stdint.h>
 #include <Arduino.h>
 
 class HCSR{
   private:
-    unsigned short trig;
-    unsigned short echo;
+    uint8_t trig;
+    uint8_t echo;
   public:
-    HCSR(unsigned short, unsigned short);
-    long read();
+    HCSR(uint8_t, uint8_t);
+    uint8_t read();
 };
 
 #endif
